0825 - 확진, 밀집 지도 DB 연결 완료

+++ src/main/java/com/megait/nocoronazone/domain/DetailSafetyIndex.java
@@ -17,5 +17,5 @@ public class DetailSafetyIndex {
     @Id
     private String district;
 
-    private int index;
+    private double index;
 }
